git-svn-id: svn://svn.r-forge.r-project.org/svnroot/bigmemory@350 edb9625f-4e0d-4859-8d74-9fd3b1da38cb

+++ pkg/bigtabulate/src/bigtabulate.cpp
@@ -680,7 +680,8 @@ SEXP TAPPLY( MatrixAccessorType m, SEXP columns, SEXP breakSexp,
         Indices &inds = tis[i];
         for (j=0; j < static_cast<index_type>(inds.size()); ++j)
         {
-          SET_STRING_ELT(mn, inds[j]-1, mkChar(groupNames[i].c_str()));
+          SET_STRING_ELT(mn, static_cast<int>(inds[j]-1), 
+            mkChar(groupNames[i].c_str()));
           pmr[static_cast<index_type>(inds[j])-1] = i+1;
         }
       }
